% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdenticon_npm_install.R
\name{jdenticon_npm_install}
\alias{jdenticon_npm_install}
\title{Install jdenticon npm dependency.}
\usage{
jdenticon_npm_install(force = FALSE)
}
\arguments{
\item{force}{\code{boolean} If \code{TRUE}, automatically proceeds with npm installation. If \code{FALSE} (default), prompts user before proceeding with npm installation.}
}
\value{
Updated \code{inst/node_modules} folder.
}
\description{
Install jdenticon npm dependency.
}
\examples{
\dontrun{
jdenticon_npm_install(force = TRUE)
}

}
